requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
pinecone-client==2.2.4
google-generativeai==0.3.2
psycopg2-binary==2.9.9
redis==5.0.1
python-multipart==0.0.6
python-dotenv==1.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3